Move http constants from lib/services/http/constants.js to lib/http/constants.js
