+++ lib/http/constants.js
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+var version = require('util/version');
+
 /**
  * The maximum number of requests client can send to the /ca/ endpoint before
  * blocking it.
@@ -60,6 +62,40 @@ var LOG_FORMAT = ':remote-addr - - [:date] :method :url HTTP/:http-version " :st
 var LIMITER_RULES = [
   [CA_REQUEST_PATH_RE, CA_REQUEST_METHOD, REQUEST_COUNT_LIMIT, REQUEST_TIME_LIMIT, true]
 ];
+
+/**
+ * List of enabled http endpoints
+ * @const
+ * @type {Array}
+ */
+var APPS = ['info', 'api', 'ca', 'bundles', 'services', 'facts', 'health', 'instances'];
+
+/**
+ * Default http services root path
+ * @const
+ * @type {String}
+ */
+var SERVICES_HTTP_ROOT = 'services/http/';
+
+/**
+ * Currently active API version
+ * @const
+ * @type {String}
+ */
+var CURRENT_API_VERSION = '1.0';
+
+/**
+ * Headers which are added to every HTTP response.
+ * @const
+ * @type {Object}
+ */
+var BASE_HEADERS = { 'X-Cast-Agent-Version': version.toString(),
+                     'X-Cast-API-Version': CURRENT_API_VERSION };
+
+exports.BASE_HEADERS = BASE_HEADERS;
+exports.APPS = APPS;
+exports.SERVICES_HTTP_ROOT = SERVICES_HTTP_ROOT;
+exports.CURRENT_API_VERSION = CURRENT_API_VERSION;
 
 exports.REQUEST_COUNT_LIMIT = REQUEST_COUNT_LIMIT;
 exports.REQUEST_TIME_LIMIT = REQUEST_TIME_LIMIT;
